Create php code to take input from input field
This code is a simple script that allows the user to download a remote file and save it locally on their [laptop](https://geeksvela.com/). It takes the input from an input field and validates it as a valid URL. Then, it uses cURL to download the file and save it in a downloads folder with the same name as the original file. It also checks if the file was downloaded successfully and displays a message accordingly. This code can be useful for downloading files from the web without using a browser.